fix: use uint64_t for chat tab typeMask to avoid UB for ChatType values >= 32, add missing boss/party chat types to Combat tab

+++ include/ui/game_screen.hpp
@@ -66,7 +66,7 @@ private:
     int activeChatTab_ = 0;
     struct ChatTab {
         std::string name;
-        uint32_t typeMask;  // bitmask of ChatType values to show
+        uint64_t typeMask;  // bitmask of ChatType values to show (64-bit: types go up to 84)
     };
     std::vector<ChatTab> chatTabs_;
     void initChatTabs();
